Refactor; page numbers in pdf view
