+++ peachjam/analysis/citations.py
@@ -1,10 +1,67 @@
 import shutil
 import tempfile
+from dataclasses import dataclass
 
 import lxml.html
 from docpipe.pdf import pdf_to_text
 
+from peachjam.analysis.matchers import TextPatternMatcher
 from peachjam.models import CitationLink
+
+
+@dataclass
+class ExtractedCitation:
+    text: str
+    start: int
+    end: int
+    href: str
+    target_id: str
+
+
+class CitationMatcher(TextPatternMatcher):
+    """Marks references to cited documents that follow a common citation pattern."""
+
+    marker_tag = "a"
+
+    href_pattern = "/akn/"
+
+    def setup(self, *args, **kwargs):
+        super().setup(*args, **kwargs)
+        self.citations = []
+
+    def handle_text_match(self, text, match):
+        self.citations.append(
+            ExtractedCitation(
+                match.group(),
+                match.start(),
+                match.end(),
+                self.make_href(match),
+                self.pagenum,
+            )
+        )
+
+    def is_node_match_valid(self, node, match):
+        if self.make_href(match) != self.frbr_uri.work_uri():
+            return True
+
+    def markup_node_match(self, node, match):
+        """Markup the match with a ref tag. The first group in the match is substituted with the ref."""
+        node, start, end = super().markup_node_match(node, match)
+        href = self.make_href(match)
+        node.set("href", href)
+        self.citations.append(
+            ExtractedCitation(match.group(), match.start(), match.end(), href)
+        )
+        return node, start, end
+
+    def make_href(self, match):
+        """Turn this match into a full FRBR URI href using the href_pattern. Subclasses can also
+        override this method to do more complex things.
+        """
+        return self.href_pattern.format(**self.href_pattern_args(match))
+
+    def href_pattern_args(self, match):
+        return match.groupdict()
 
 
 class CitationAnalyser:
@@ -37,7 +94,8 @@ class CitationAnalyser:
             # convert document to pdf and then extract the text
             pdf = document.source_file.as_pdf()
             shutil.copyfileobj(pdf, tmp)
-            text = pdf_to_text(tmp)
+            tmp.flush()
+            text = pdf_to_text(tmp.name)
 
         for matcher in self.matchers:
             matcher = matcher()
@@ -46,8 +104,10 @@ class CitationAnalyser:
 
     def store_text_citation_links(self, document, matcher):
         """Transform extracted citations from text into CitationLink objects."""
-        for citation in matcher.citations:
-            document.citation_links.add(CitationLink.from_extracted_citation(citation))
+        citations = [CitationLink.from_extracted_citation(c) for c in matcher.citations]
+        for c in citations:
+            c.document = document
+        document.citation_links.add(*citations, bulk=False)
 
 
 citation_analyser = CitationAnalyser()
