+++ peachjam/analysis/matchers.py
@@ -1,5 +1,3 @@
-from dataclasses import dataclass
-
 from lxml import etree
 
 from peachjam.analysis.xmlutils import wrap_text
@@ -49,7 +47,7 @@ class TextPatternMatcher:
 
     def extract_paged_text_matches(self):
         # split on form feed (page break)
-        for i, page in enumerate(self.text.split("\0xC")):
+        for i, page in enumerate(self.text.split("\x0C")):
             self.pagenum = i
             self.run_text_extraction(page)
 
@@ -132,58 +130,3 @@ class TextPatternMatcher:
 
     def candidate_text_nodes(self, root):
         return self.candidate_xpath(root)
-
-
-@dataclass
-class ExtractedCitation:
-    text: str
-    start: int
-    end: int
-    href: str
-    target_id: str
-
-
-class CitationMatcher(TextPatternMatcher):
-    """Marks references to cited documents that follow a common citation pattern."""
-
-    marker_tag = "a"
-
-    href_pattern = "/akn/"
-
-    def setup(self, *args, **kwargs):
-        super().setup(*args, **kwargs)
-        self.citations = []
-
-    def handle_text_match(self, text, match):
-        self.citations.append(
-            ExtractedCitation(
-                match.group(),
-                match.start(),
-                match.end(),
-                self.make_href(match),
-                self.pagenum,
-            )
-        )
-
-    def is_node_match_valid(self, node, match):
-        if self.make_href(match) != self.frbr_uri.work_uri():
-            return True
-
-    def markup_node_match(self, node, match):
-        """Markup the match with a ref tag. The first group in the match is substituted with the ref."""
-        node, start, end = super().markup_node_match(node, match)
-        href = self.make_href(match)
-        node.set("href", href)
-        self.citations.append(
-            ExtractedCitation(match.group(), match.start(), match.end(), href)
-        )
-        return node, start, end
-
-    def make_href(self, match):
-        """Turn this match into a full FRBR URI href using the href_pattern. Subclasses can also
-        override this method to do more complex things.
-        """
-        return self.href_pattern.format(**self.href_pattern_args(match))
-
-    def href_pattern_args(self, match):
-        return match.groupdict()
